# Optimized performances and improved code readability