--- a/common/src/main/java/net/threetag/palladium/power/PowerManager.java
+++ b/common/src/main/java/net/threetag/palladium/power/PowerManager.java
@@ -34,6 +34,7 @@ public class PowerManager extends SimpleJsonResourceReloadListener {
     private static PowerManager INSTANCE;
     public Map<ResourceLocation, Power> byName = ImmutableMap.of();
     public static final List<String> CHECK_FOR_CHAT_MESSAGES = new ArrayList<>();
+    private static final Map<PowerHandler, Optional<PowerHandler>> POWER_HANDLER_MAP = new WeakHashMap<>();
 
     public static void init() {
         ReloadListenerRegistry.register(PackType.SERVER_DATA, Palladium.id("powers"), INSTANCE = new PowerManager());
@@ -102,10 +103,15 @@ public class PowerManager extends SimpleJsonResourceReloadListener {
     }
 
     public static Optional<PowerHandler> getPowerHandler(LivingEntity entity) {
-        if (entity instanceof PalladiumLivingEntityExtension ext) {
-            return Optional.of(ext.palladium$getPowerHandler());
-        } else {
+        if (!(entity instanceof PalladiumLivingEntityExtension ext)) {
             return Optional.empty();
         }
+
+        final PowerHandler handler = ext.palladium$getPowerHandler();
+        if (handler == null) {
+            return Optional.empty();
+        }
+
+        return POWER_HANDLER_MAP.computeIfAbsent(handler, Optional::of);
     }
 }