Hello,

Performance improvement for power and abilities. It results in lower server resources usage, MSPT and CPU usage. Plugin itself runs ~2.44x faster

Here's some comparison stats from spark profiler, both ran for 15 minutes:
Base:
<img width="1248" height="92" alt="image" src="https://github.com/user-attachments/assets/19a4fd72-111a-42e7-825a-0365185f8344" />
Optimized:
<img width="1247" height="97" alt="image" src="https://github.com/user-attachments/assets/2e0da474-d96e-4325-ad0c-ed6e69085e58" />


**Time per tick:**
Base:
- Palladium: 8.33ms
- PalladumCore: 7.54ms
- Total: 15.87ms

Optimized:
- Palladium: 3.57ms
- PalladiumCore: 2.93ms
- Total: 6.5ms