@DeveloperAnonymous 2025-08-05
Ohh ok, do you want to keep it like that or should I find a way to make it work properly? Caching gave some performance improvements